requirements.txt
llama-cpp-python
requests
rich